feat: shift table component
